+++ baby-gru/package.json
@@ -1,5 +1,5 @@
 {
-  "name": "Moorhen",
+  "name": "moorhen",
   "version": "0.1.0",
   "private": true,
   "main": "public/electron.js",
@@ -82,6 +82,7 @@
   },
   "scripts": {
     "start": "react-scripts start",
+    "start-tsc": "npx nodemon --trace-warnings",
     "build": "react-scripts build",
     "build-cloud-dev": "webpack  --mode development --devtool inline-source-map",
     "build-cloud": "webpack --mode production",
@@ -120,7 +121,8 @@
     "@electron-forge/maker-rpm": "^6.0.5",
     "@electron-forge/maker-squirrel": "^6.0.5",
     "@electron-forge/maker-zip": "^6.0.5",
-    "@types/node": "^18.11.9",
+    "@types/node": "^18.16.14",
+    "@types/react": "^18.2.6",
     "babel-core": "^6.26.3",
     "babel-loader": "^9.1.0",
     "babel-preset-env": "^1.7.0",
@@ -130,7 +132,11 @@
     "electron": "^23.1.3",
     "electron-devtools-installer": "^3.2.0",
     "html-webpack-plugin": "^5.5.0",
+    "nodemon": "^2.0.22",
     "style-loader": "^3.3.1",
+    "ts-loader": "^9.4.3",
+    "ts-node": "^10.9.1",
+    "typescript": "^5.0.4",
     "webpack": "^5.75.0",
     "webpack-cli": "^5.0.1",
     "webpack-dev-server": "^4.11.1",
